% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{add.id.2}
\alias{add.id.2}
\title{Add region ID}
\usage{
add.id.2(ranges, reg.length = 3000)
}
\arguments{
\item{ranges}{GRanges object}

\item{reg.length}{length of regions}
}
\value{
GRanges object with ID column
}
\description{
Adds column containing region id (chr:start:end) to every track line in base resolution GRanges object,
so that adjacent positions in the same region get the same ID. 
All ranges must have the same length.
}
