% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{all.hexamers}
\alias{all.hexamers}
\title{All possible hexamers}
\usage{
\method{all}{hexamers}()
}
\value{
vector containing all possible DNA hexamer sequences (permutations of 4 letters = 4096 hexamers)
}
\description{
All possible hexamers
}
