% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{batch.prob.uniform}
\alias{batch.prob.uniform}
\title{Primer probability from batch}
\usage{
batch.prob.uniform(hexs = all.hexamers(), nuc.probs = c(pA = 0.25, pT =
  0.25, pG = 0.25, pC = 0.25))
}
\arguments{
\item{hexs}{vector of random hexamer sequences}

\item{nuc.probs}{vector of fraction of each nucleotide}
}
\value{
vector of primer probabilities
}
\description{
returns probability of having each possible hexamer primer ASSUMING THE SAME NUCLEOTIDE COMPOSITION IN EVERY POSITION OF THE HEXAMER SEQUENCE
}
