% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{build.random.hex}
\alias{build.random.hex}
\title{Random hexamer}
\usage{
build.random.hex(pA = 0.25, pT = 0.25, pG = 0.25, pC = 0.25)
}
\arguments{
\item{pA}{fraction of A}

\item{pT}{fraction of T}

\item{pG}{fraction of G}

\item{pC}{fraction of C}
}
\value{
string of random hexamer
}
\description{
Generates a random hexamer pased on input sequence composition
}
