% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{compare.spear.real}
\alias{compare.spear.real}
\title{Quantify correlation of predicted and experimental coverage}
\usage{
compare.spear.real(track.list, name.real = "WGS", method = "spearman")
}
\arguments{
\item{track.list}{GRangesList object of sampled regions with experimental and predicted coverage}

\item{name.real}{character, name of seq library}

\item{method}{a character string indicating which correlation coefficient is to be computed. One of "pearson", "kendall", or "spearman" (default)}
}
\value{
list of vector for real correlation coefficients, for coefficients of random permutation, ggplot object of boxplot
}
\description{
For a given GRangesList, computes correlation between pred and exp profiles in every regions, in randomnly perputed regions 
and plots boxplot of comparisons, with p-value for Wilcoxon's test.
}
