% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{compute.keqs}
\alias{compute.keqs}
\title{Compute association constants}
\usage{
\method{compute}{keqs}(pt.df, eps, filter.pt = 200)
}
\arguments{
\item{pt.df}{long dataframe of primer-template features}

\item{eps}{epsilon value for the sample}

\item{filter.pt}{integer. The minimum number of pt counts to keep a primer template pair}
}
\value{
numeric of epsilon value (minimum of Chi-sq. function)
}
\description{
Compute K with model parameters
}
