% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{compute.primer.usage}
\alias{compute.primer.usage}
\title{Compute primer usage}
\usage{
\method{compute}{primer.usage}(pt.all.df)
}
\arguments{
\item{pt.all.df}{long dataframe of primer-template counts (matching and mismatching)}
}
\value{
long dataframe of primer-template features
}
\description{
Compute primer usage
}
