% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{correlate.pred.exp}
\alias{correlate.pred.exp}
\title{Correlation of predicted and experimental coverage}
\usage{
correlate.pred.exp(pred, exp, method = "spearman")
}
\arguments{
\item{pred}{vector of predicted coverage profile}

\item{exp}{vector of experimental coverage profile}

\item{method}{a character string indicating which correlation coefficient is to be computed. One of "pearson", "kendall", or "spearman" (default)}
}
\value{
correlation coefficient
}
\description{
Correlation of predicted and experimental coverage
}
