% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{coverage.yield.delta}
\alias{coverage.yield.delta}
\title{Compute difference in yield between batches}
\usage{
coverage.yield.delta(scaled.track, roi.track)
}
\description{
Computes difference in coverage yield for region of interest between predicted coverage profiles
of two different primer batches
}
\details{
@param scaled.track GRanges object with values for predicted coverage of best batch (best) and random batch (even) (scaled, so no values < 0)
 @param roi.track GRanges object of regions of interest
 
 @return difference in yield
 
 @export
}
