% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{coverage.yield.single}
\alias{coverage.yield.single}
\title{Compute coverage yield}
\usage{
coverage.yield.single(scaled.track, roi.track)
}
\arguments{
\item{scaled.track}{GRanges object with values for predicted coverage of best batch (best) and random batch (even) (scaled, so no values < 0)}

\item{roi.track}{GRanges object of regions of interest}
}
\value{
score for coverage yield
}
\description{
Computes ratio between area under the curve in region of interest and outside of region of interest 
(sample of the same bps of the region of interest)
}
