% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{delta.yield.permutation}
\alias{delta.yield.permutation}
\title{Compute difference in yield of permuted coverage profile}
\usage{
delta.yield.permutation(my.track, roi.track, permute = "best",
  threads = detectCores())
}
\description{
Compute difference in yield of permuted coverage profile
}
