% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{epsilon.minimize.chisq}
\alias{epsilon.minimize.chisq}
\title{Max. likelihood estimation of scaling factor epsilon}
\usage{
epsilon.minimize.chisq(pt.df, max, min = 0, prob = batch.prob.uniform(),
  plot = T)
}
\arguments{
\item{pt.df}{long dataframe of primer-template features}

\item{max}{maximum epsilon value}

\item{min}{minimum epsilon value}

\item{prob}{dataframe of primer probability (default is equal for all)}

\item{plot}{logical indicating whether to plot the Chi-square function (to visualize the minimum)}
}
\value{
numeric of epsilon value (minimum of Chi-sq. function)
}
\description{
From primer-binding model parameters find epsilon value that minimizes the difference between predicted K and K as defined from calorimetry experiments.
}
