% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{find.rev.pairs}
\alias{find.rev.pairs}
\title{Switch primer and template}
\usage{
find.rev.pairs(df, smp)
}
\arguments{
\item{df}{dataframe of primer-template features}

\item{smp}{string of template sequence}
}
\value{
dataframe of selected primer-template pair
}
\description{
Finds reverse complementary of template sequence to pair primer-template pairs that are supposed to have the same binding energy
}
