% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{get.avg.methylation}
\alias{get.avg.methylation}
\title{Compute average methylation of genomic region}
\usage{
get.avg.methylation(reg.track, baseres.met)
}
\arguments{
\item{reg.track}{GRanges track of one sampled region (single element of GRanges List slit by ID)}

\item{baseres.met}{GRanges object of base resolution methylation fraction}
}
\value{
average meth fraction in input region
}
\description{
Compute average methylation of genomic region
}
