% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{get.range.methylation}
\alias{get.range.methylation}
\title{Extract methylation of genomic region}
\usage{
get.range.methylation(reg.track, baseres.met)
}
\arguments{
\item{reg.track}{GRanges track of one sampled region (single element of GRanges List slit by ID)}

\item{baseres.met}{GRanges object of base resolution methylation fraction}
}
\value{
GRanges object of methylation values over the input region
}
\description{
Extract methylation of genomic region
}
