% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{hexamerMatrix}
\alias{hexamerMatrix}
\title{Compute all possible nucleotide compositions}
\usage{
hexamerMatrix(stepSize = 0.1)
}
\arguments{
\item{stepSize}{Minimum difference in nucleotide fraction (the smaller, the more combinations)}
}
\value{
Matrix of all possible combinations of nucleotide compositions
}
\description{
Computes all possible probability combinations for 4 nucletides given a step size for the difference in probability
}
