% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{is.palindrome}
\alias{is.palindrome}
\title{Check for palindrome sequences}
\usage{
is.palindrome(seq)
}
\arguments{
\item{seq}{strind of DNA sequence}
}
\value{
logical indicating wheather the sequence is a palindrome or not
}
\description{
checks is DNA sequence is a palindrome
}
