% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{join.pt.data}
\alias{join.pt.data}
\title{Build dataframe of primer-template data}
\usage{
join.pt.data(pt.tab, template.usage, kmer.ab, tabDg)
}
\arguments{
\item{pt.tab}{long dataframe of primer-template counts}

\item{template.usage}{dataframe of template usage}

\item{kmer.ab}{dataframe of kmer genomic abundance}

\item{tabDg}{dataframe of NN delta Gs}
}
\value{
long dataframe of primer-template features
}
\description{
Build dataframe of primer-template data
}
