% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{load.expVSpred.coverage}
\alias{load.expVSpred.coverage}
\title{Load coverage profiles}
\usage{
load.expVSpred.coverage(pred.bw.file, exp.bw.file)
}
\arguments{
\item{pred.bw.file}{path to bigWig file of predicted coverage profile}

\item{exp.bw.file}{path to bigWig file of experimental coverage profile}
}
\value{
list containing 2 GRanges objects
}
\description{
Loads experimental and predicted coverage profiles from bigWig files
}
