% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{load.kmer.abundance}
\alias{load.kmer.abundance}
\title{Load genomic kmer abundance}
\usage{
load.kmer.abundance(kmer.abundance.file)
}
\arguments{
\item{kmer.abundance.file}{path to csv file of kmer abundance}
}
\value{
dataframe of kmer abundance
}
\description{
Load genomic kmer abundance
}
