% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{load.modelled.deltaG}
\alias{load.modelled.deltaG}
\title{Load NN deltaG values}
\usage{
load.modelled.deltaG(deltaG.file)
}
\arguments{
\item{deltaG.file}{path to csv file of NN delta G}
}
\value{
dataframe of NN delta G
}
\description{
Load NN deltaG values
}
