% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{load.pt.data}
\alias{load.pt.data}
\title{Load primer-template matrix}
\usage{
load.pt.data(ptCounts.file, diag.pairs = F)
}
\arguments{
\item{ptCounts.file}{path to csv file of p-t matrix}

\item{diag.pairs}{logical indicating whether to retain only diagonal pairs (matching primer-template pairs)}
}
\value{
long dataframe of primer-template counts and usage
}
\description{
Loads csv file of primer-template matrix, computing primer and template usage.
}
