% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{loadPtMatrix}
\alias{loadPtMatrix}
\title{Load primer-template matrix}
\usage{
loadPtMatrix(file, compression = "gzip")
}
\arguments{
\item{compression}{compression type of csv file ('none' for no compression)}

\item{ptCounts.file}{path to csv file of p-t matrix}
}
\value{
long dataframe of primer-template counts
}
\description{
Loads csv file of primer-template matrix
}
