% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{make.base.res.bw}
\alias{make.base.res.bw}
\title{Genomic ranges to base resolution}
\usage{
make.base.res.bw(bw)
}
\arguments{
\item{bw}{GRanges object}
}
\value{
GRanges object of base resolution track
}
\description{
Breaks GRanges object to single base resolution, giving same score to every base from a same region
}
