% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{make.match.df}
\alias{make.match.df}
\title{Build dataframe of matching primer-template data}
\usage{
make.match.df(pt.file, ab.file)
}
\arguments{
\item{pt.file}{path to csv file of primer-template counts}

\item{ab.file}{path to csv file of kmer genomic abundance}
}
\value{
long dataframe of matching primer-template features
}
\description{
Build dataframe of matching primer-template data
}
