% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{make.predVSexp.range}
\alias{make.predVSexp.range}
\title{Merge coverage profiles}
\usage{
make.predVSexp.range(pred.bw, exp.bw, pred.name = "pred",
  exp.name = "score")
}
\arguments{
\item{pred.bw}{GRanges track (predicted coverage profile)}

\item{exp.bw}{GRanges track (experimental coverage profile)}

\item{pred.name}{character of name for first track score column (e.g. "predicted")}

\item{exp.name}{character of name for second track score column}
}
\value{
GRanges track of common ranges with 2 score columns
}
\description{
Generates a single GRanges object containing overlapping ranges for 2 GRanges objects with score (coverage)
}
