% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{make.predVSexp.track}
\alias{make.predVSexp.track}
\title{Make track of predicted VS experimental coverage}
\usage{
make.predVSexp.track(predicted.bw, experimental.bw, reg.length = 3000)
}
\arguments{
\item{predicted.bw}{GRanges object of predicted coverage track}

\item{experimental.bw}{GRanges object of experimental coverage track}

\item{reg.length}{length of sampled regions}
}
\value{
GRangesList object where each element is a sampled region
}
\description{
Merges, smoothens and normalizes coverage tracks for experimental and predicted coverage, adding IDs for each sampled region
(region length has to be specified)
}
