% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{make.random.profile}
\alias{make.random.profile}
\title{Make random permutation of coverage profile}
\usage{
make.random.profile(cum.dist.real, threads = detectCores())
}
\arguments{
\item{cum.dist.real}{cumulative distribution of coverage scores}

\item{threads}{amound of threads to use}
}
\value{
vector of permuted coverage profile
}
\description{
To calculate p-value of difference in yield
}
