% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{make_pair_df}
\alias{make_pair_df}
\title{Make long dataframe of primer-template counts}
\usage{
make_pair_df(dgMat)
}
\arguments{
\item{dgMat}{matrix of primer-template counts}
}
\value{
long dataframe of primer-template counts
}
\description{
Make long dataframe of primer-template counts
}
