% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{make_ppm_of_usage}
\alias{make_ppm_of_usage}
\title{From primer usage to position probability matrix}
\usage{
make_ppm_of_usage(primer.usage.df)
}
\arguments{
\item{primer.usage.df}{dataframe of primer usage}
}
\value{
position probability matrix of nucleotide composition of primers
}
\description{
Computes matrix of nucleotide composotion of primers
}
