% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{nice.plotTrack}
\alias{nice.plotTrack}
\title{Plot comparison of coverage profiles}
\usage{
nice.plotTrack(region.track, labels = c("Experimental ", "Predicted"))
}
\arguments{
\item{region.track}{GRanges track of one sampled region (single element of GRanges List slit by ID)}

\item{labels}{vector of names for the 2 compared coverage profiles}
}
\value{
ggplot object
}
\description{
Make ggplot viz of predicted and experimental coverage tracks for a sampled region of the genome
}
