% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{norm.scale.n.yield}
\alias{norm.scale.n.yield}
\title{Normalize and compute yield}
\usage{
norm.scale.n.yield(raw.track, roi.track, scale.by = 5)
}
\arguments{
\item{raw.track}{GRanges object with values for predicted coverage of best batch (best) and random batch (even)}

\item{roi.track}{GRanges object of regions of interest}

\item{scale.by}{scaling fator to add to each coverage value to have all > 0}
}
\value{
score for coverage yield
}
\description{
Computes ratio between area under the curve in region of interest and outside of region of interest 
(sample of the same bps of the region of interest) after normalization and scaling to have no coverage values < 0
}
