% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{normalize.coverage}
\alias{normalize.coverage}
\title{Standard normalize coverage}
\usage{
normalize.coverage(bw)
}
\arguments{
\item{bw}{GRanges object with score columns}
}
\value{
GRanges object with normalized coverage profiles
}
\description{
Compute z-score normalization (x-mean(x))/sd(x)) for coverage profiles
}
