% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{plot.batch.accuracy}
\alias{plot.batch.accuracy}
\title{Plot accuracy of prediction for variable primer concentrations}
\usage{
plot.batch.accuracy(pcc.primer.batch)
}
\arguments{
\item{pcc.primer.batch}{dataframe of R sq. values for prediction with different primer batches (output of \code{run_cov_prediction.r} script)}
}
\value{
plot for nucleotide concentration VS R.squared
}
\description{
Plot accuracy of prediction for variable primer concentrations
}
