% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{plot.expVSpred.coverage.track}
\alias{plot.expVSpred.coverage.track}
\title{Plot predicted and experimental coverage profiles}
\usage{
plot.expVSpred.coverage.track(common.bw, met = NULL, plot = TRUE)
}
\description{
Deprecated, use niceplotTrack
}
