% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{plot.prediction}
\alias{plot.prediction}
\title{Plot predicted coverage VS experimental}
\usage{
plot.prediction(pred.cov.df, color = "nuc")
}
\arguments{
\item{pred.cov.df}{dataframe of primer-template features with column for predicted coverage}

\item{color}{feature to pass to color aesthetic: "nuc" (default) colors by prevalent nucleotide in template sequence, "CG" colors by present or absence of CG dinucleotide in template sequence}
}
\value{
scatterplot
}
\description{
Make scatterplot of predicted and experimental template coverage fraction
}
