% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{predict.coverage}
\alias{predict.coverage}
\title{Coverage prediction}
\usage{
\method{predict}{coverage}(keqs.df, eps, prob = batch.prob.uniform())
}
\arguments{
\item{keqs.df}{long dataframe of primer-template features, including association constants for primer-template binding reactions}

\item{eps}{epsilon value for the sample}

\item{prob}{dataframe of primer probability (default is equal for all)}
}
\value{
dataframe of primer-template features with column for predicted coverage
}
\description{
Predict template sequence coverage based on primer binding model
}
