% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{prevalent_nucleotide}
\alias{prevalent_nucleotide}
\title{Compute most abundant nucleotide in sequence}
\usage{
prevalent_nucleotide(seq)
}
\arguments{
\item{seq}{string of nucleotide sequence}
}
\value{
string of most abundant nucleotide
}
\description{
doesn't deal with ties
}
