% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{primer.prob}
\alias{primer.prob}
\title{Primer probability}
\usage{
primer.prob(seq, probs = c(pA = 0.25, pT = 0.25, pG = 0.25, pC = 0.25))
}
\arguments{
\item{seq}{string of DNA sequence}

\item{pA}{fraction of A}

\item{pT}{fraction of T}

\item{pG}{fraction of G}

\item{pC}{fraction of C}
}
\value{
numerical of probability
}
\description{
Computes the probability of having a given sequence in the primer pool based on input nucleotide composition of primers
}
