% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{random.delta.yield.dist}
\alias{random.delta.yield.dist}
\title{Random yield}
\usage{
random.delta.yield.dist(my.track, roi.track, n.iterations = 1000,
  threads = detectCores(), verbose = T)
}
\arguments{
\item{my.track}{GRanges object with values for predicted coverage of best batch (best) and random batch (even) (SCALED so no cov < 0)}

\item{roi.track}{GRanges object of regions of interest}

\item{n.iterations}{number of iterations}

\item{threads}{number of threads to use}

\item{verbose}{logical indicating wheather to print progress}
}
\value{
list of delta yield and vector of delta yield for random permutations
}
\description{
Computes difference in coverage yield for region of interest between predicted coverage profiles
of two different primer batches. To calculate p-value of yield score, it also computes the distribution of 
difference in yield between randomnly shuffled coverage profiles.
}
