% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{random.from.even.2}
\alias{random.from.even.2}
\title{Make random permutation of coverage profile for random batch}
\usage{
random.from.even.2(vec, vec.cum.dist)
}
\description{
Make random permutation of coverage profile for random batch
}
