% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{randomize}
\alias{randomize}
\title{Compute coverage correlation of random permutations of predicted coverage}
\usage{
randomize(track.list, method = "spearman")
}
\arguments{
\item{track.list}{GRangesList object of sampled regions with experimental and predicted coverage}

\item{method}{a character string indicating which correlation coefficient is to be computed. One of "pearson", "kendall", or "spearman" (default)}
}
\value{
vector of correlation coefficients of permutations of sampled regions
}
\description{
Given a GRangesList of sampled regions, where for each sample the experimental and predicted coverage profile

is given, the function shuffles the predicted coverage profiles, assigning them to random regions and computes the correlation coefficient
}
