% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{rev.comp}
\alias{rev.comp}
\title{Reverse and/or complement sequence}
\usage{
\method{rev}{comp}(x, compl = TRUE, rev = TRUE)
}
\arguments{
\item{x}{string of DNA sequence}

\item{compl}{logical indicating wheather to take the complement of sequence x}

\item{rev}{logical indicating wheather to take the reverse of sequence x}
}
\value{
string of transformed sequence
}
\description{
Reverse and/or complement sequence
}
