% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binding_model_functions.r
\name{simulate.primer.pool}
\alias{simulate.primer.pool}
\title{Primer pool simulation}
\usage{
\method{simulate}{primer.pool}(pool.size = 50000, pA = 0.25, pT = 0.25,
  pG = 0.25, pC = 0.25)
}
\arguments{
\item{pool.size}{integer, number of hexamer sequences in the pool (the more the longer it takes)}

\item{pA}{fraction of A}

\item{pT}{fraction of T}

\item{pG}{fraction of G}

\item{pC}{fraction of C}
}
\value{
vector containing the simulated random hexamers
}
\description{
Generates a pool of random hexamers based on input nucleotide composition
}
