% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{split.tracks}
\alias{split.tracks}
\title{Split ranges by ID}
\usage{
\method{split}{tracks}(range, reg.length)
}
\arguments{
\item{range}{GRanges object}

\item{reg.length}{length of regions}
}
\value{
GRangesList object as long as the number of regions
}
\description{
Splits GRanges object in GRanges list of ranges with the same ID
}
