% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{subsetByRegion}
\alias{subsetByRegion}
\title{Subset by region}
\usage{
subsetByRegion(ranges, chrom, start, end)
}
\arguments{
\item{ranges}{GRanges object}

\item{chrom}{selected chromosome}

\item{start}{selected starting position}

\item{end}{selected ending position}
}
\value{
GRanges object of track in selected coordinates
}
\description{
Subset by region
}
