% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_peaks.r
\name{test.add.id}
\alias{test.add.id}
\title{Test IDs}
\usage{
test.add.id(range, reg.length = 3000)
}
\arguments{
\item{range}{GRanges object with id column}

\item{reg.length}{length of regions}
}
\value{
'good' or 'bad'
}
\description{
Tests if region IDs correspond to start and end of region
}
